Constrain target of addrof operator to PTR
Adds an edge PTR -> x for expressions &x. Previously, only the result
of the addrof expression had this constraint, but CheckedC prohibits
taking the address of pointers with bounds. It is possible to take the
address of an _Array_ptr when the pointer doesn't have bounds declared.
With this constraint added however, 3C will not let this happen.