Fix function rewriting with r-paren between prototype and definition
A right parentheses appearing between a function prototype and the
beginning of its definition caused anything falling between the
parenthesis and the actual end of the prototype to be deleted. This
problem is fixed by using clang provided source locations instead of
locating the closing parenthesis by iterating of the source code
characters.