More consistent handling of rewriting around macros (fix #429)
Previously a macro appearing at the beginning of rewritten source range
would not cause failure in rewriting, but a macro coming at the end of
the range would. This was because the "expansion" source location was
used for the start of the range, but it was not used for the end.
With this change, we also take the expansion of the end of the range.

In some parts of the rewriter, the expansion location was not used for
either end of the range. These have been changed to use the same logic
as the other rewritings. This change will allow some rewritings that
would have been silently dropped and emit an error on others where no
rewriting is possible.

Since using expansion locations is likely to clobber macros in the
source code and replace them with their expansion, a warning is now
emitted when the rewriter uses the expansion locations instead of the
original locations.