Generate constraints so that casts are not required inside macros
The constraint created between a function parameter and a corresponding
argument is updated to be an equality constraint when the function call
is inside a macro. Doing this forces the checked type of the parameter
to the same as that of the argument. Casts are only inserted when
parameter and argument types are not equal, so this changes ensures that
3C will not try (and subsequently fail) to insert casts on expressions
in macros.