Remove the DComp custom comparator for the rewrite set (RSet).
3C may have relied on this custom order at one time, but it no longer
seems to; at least, the regression tests pass after this change.
Rewriting of non-overlapping declarations should be independent, and the
code already ensures that all rewrites to a given multi-decl are
performed together in the appropriate way.

RSet is now a std::map<Decl *, DeclReplacement *> so that
rewriteMultiDecl can look up the rewrite (if any) for a given Decl in
the multi-decl rather than scanning the set in the original custom
order. The use of a map adds some boilerplate, but I can't think of a
better option.

Also, VisitedMultiDeclMembers can just be a std::set<Decl *> because
that's all we need in order to test whether a Decl has already been
rewritten.