mkString simplifications.
As I dug into the reasons why all the conditionals in the PR so far were
needed under the current design of the rest of mkString, I realized that
the design could be straightened out a bit and correctly handle one more
case involving checked pointers to fixed-size arrays. This doesn't
address the more difficult problems with wild pointers to fixed-size
arrays or with wild pointer levels being in reverse order in general
(see https://github.com/correctcomputation/checkedc-clang/issues/703).