Fix checked pointers to arrays of wild function pointers.
Any array levels still pending when we reach a function pointer should
become part of the function pointer declarator even if some text for an
outer checked pointer level has already been added to Ss. However, any
array levels that are outside the checked pointer level and have already
been committed to EndStrs should remain at the end of the entire type
and not get sucked into the function pointer declarator. To maintain
this distinction, we need to use a separate `FptrEndStrs` list for the
array levels in the function pointer declarator.