-itypes-for-extern should not force supplementary decls to be unchecked.
This could cause a compile error if the rest of the function assumed the
variable could be used at the solved checked type of the internal
PVConstraint.

This forced me to factor out the duplicate code related to supplementary
decls from DeclRewriter::build{Itype,Checked}Decl as I had proposed.