Work around issue encounted in vsftpd port phase two
Lower bounds inference was not functioning properly when a function
declaration containing a paramter that was rewriten using a fresh lower
bound was coppied into the working directory at the start of phase two.
The paramter apears in the header with a count bound, but the fresh
lower bound does not appear in the function body because the source
file hasn't been coppied yet.

This is a hack. A better solution for this problem should be found.