Add a flag to disable inferring lower bounds.
Note that the behavior under this flag is not identical to the behavior
of 3C prior to implementing lower bounds inference. Some bounds that
were previously inferred are now not inferred because array pointers are
transitively considered invalid lower bounds due to assignments to
pointers updated in pointer arithmetic expressions, and the absences of
lower bound inference means that now lower bound will ever be inferred,
causing the rest of the bounds inference algorithm to ignore these array
pointers.